e2e: fix accesslog test
