+++ test/e2e/tests/utils.go
@@ -7,10 +7,12 @@ package tests
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
 	"net/http"
+	"net/url"
 	"strings"
 	"testing"
 	"time"
@@ -30,6 +32,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	gwapiv1a2 "sigs.k8s.io/gateway-api/apis/v1alpha2"
 	"sigs.k8s.io/gateway-api/conformance/utils/config"
+	"sigs.k8s.io/gateway-api/conformance/utils/suite"
 	"sigs.k8s.io/gateway-api/conformance/utils/tlog"
 
 	egv1a1 "github.com/envoyproxy/gateway/api/v1alpha1"
@@ -364,4 +367,97 @@ func WaitForLoadBalancerAddress(t *testing.T, client client.Client, timeout time
 	})
 	require.NoErrorf(t, waitErr, "error waiting for Service to have at least one load balancer IP address in status")
 	return ipAddr, nil
+}
+
+func ALSLogCount(t *testing.T, suite *suite.ConformanceTestSuite) int {
+	metricPath, err := RetrieveURL(suite.Client, types.NamespacedName{
+		Namespace: "monitoring",
+		Name:      "envoy-als",
+	}, 19001, "/metrics")
+	if err != nil {
+		t.Fatalf("failed to get metric url: %v", err)
+	}
+
+	countMetric, err := RetrieveMetric(metricPath, "log_count", time.Second)
+	if err != nil {
+		t.Fatalf("failed to get metric: %v", err)
+	}
+
+	total := 0
+	for _, m := range countMetric.Metric {
+		if m.Counter != nil && m.Counter.Value != nil {
+			total += int(*m.Counter.Value)
+		}
+	}
+
+	return total
+}
+
+// QueryLogCountFromLoki queries log count from loki
+func QueryLogCountFromLoki(t *testing.T, c client.Client, keyValues map[string]string, match string) (int, error) {
+	svc := corev1.Service{}
+	if err := c.Get(context.Background(), types.NamespacedName{
+		Namespace: "monitoring",
+		Name:      "loki",
+	}, &svc); err != nil {
+		return -1, err
+	}
+	lokiHost := ""
+	for _, ing := range svc.Status.LoadBalancer.Ingress {
+		if ing.IP != "" {
+			lokiHost = ing.IP
+			break
+		}
+	}
+
+	qParams := make([]string, 0, len(keyValues))
+	for k, v := range keyValues {
+		qParams = append(qParams, fmt.Sprintf("%s=\"%s\"", k, v))
+	}
+
+	q := "{" + strings.Join(qParams, ",") + "}"
+	if match != "" {
+		q = q + "|~\"" + match + "\""
+	}
+	params := url.Values{}
+	params.Add("query", q)
+	params.Add("start", fmt.Sprintf("%d", time.Now().Add(-10*time.Minute).Unix())) // query logs from last 10 minutes
+	lokiQueryURL := fmt.Sprintf("http://%s:3100/loki/api/v1/query_range?%s", lokiHost, params.Encode())
+	res, err := http.DefaultClient.Get(lokiQueryURL)
+	if err != nil {
+		return -1, err
+	}
+	t.Logf("get response from loki, query=%s, status=%s", q, res.Status)
+
+	b, err := io.ReadAll(res.Body)
+	if err != nil {
+		return -1, err
+	}
+
+	lokiResponse := &LokiQueryResponse{}
+	if err := json.Unmarshal(b, lokiResponse); err != nil {
+		return -1, err
+	}
+
+	if len(lokiResponse.Data.Result) == 0 {
+		return 0, nil
+	}
+
+	total := 0
+	for _, res := range lokiResponse.Data.Result {
+		total += len(res.Values)
+	}
+	t.Logf("get response from loki, query=%s, total=%d", q, total)
+	return total, nil
+}
+
+type LokiQueryResponse struct {
+	Status string `json:"status"`
+	Data   struct {
+		ResultType string `json:"resultType"`
+		Result     []struct {
+			Metric interface{}
+			Values []interface{} `json:"values"`
+		}
+	}
 }
